Renamed and moved treenode base.
