+++ src/crossdev64/emulator/EmulatorModuleNodes.java
@@ -3,11 +3,11 @@ package crossdev64.emulator;
 import java.awt.Window;
 
 import crossdev64.emulator.vice.VICESettingsNode;
+import crossdev64.gui.TreeNodeBase;
 import crossdev64.settings.GlobalSettings;
-import crossdev64.settings.nodes.SettingsNodeBase;
 
 public class EmulatorModuleNodes
-	extends SettingsNodeBase
+	extends TreeNodeBase
 {
 	public static final String MODULE_ID = "21453DC0-866F-433B-95D4-E25C984D9824";
 
@@ -25,7 +25,7 @@ public class EmulatorModuleNodes
 	}
 
 	@Override
-	public SettingsNodeBase createItem(Window oParent, SettingsNodeBase oDefault)
+	public TreeNodeBase createItem(Window oParent, TreeNodeBase oDefault)
 	{
 		// TODO: For now we just create a VICE node. If we support other emulators as well, we need to replace it with a selection dialog. 
 		return new VICESettingsNode();
