+++ src/crossdev64/settings/SettingsTreeModel.java
@@ -2,25 +2,25 @@ package crossdev64.settings;
 
 import javax.swing.tree.DefaultTreeModel;
 
-import crossdev64.settings.nodes.SettingsNodeBase;
+import crossdev64.gui.TreeNodeBase;
 
 public class SettingsTreeModel
 	extends DefaultTreeModel
 {
 	private static final long serialVersionUID = 1L;
 
-	private SettingsNodeBase mRoot;
+	private TreeNodeBase mRoot;
 	
 	public SettingsTreeModel()
 	{
 		super(GlobalSettings.getInstance().getRootNode());
-		mRoot = (SettingsNodeBase)super.getRoot();
+		mRoot = (TreeNodeBase)super.getRoot();
 		initModules(mRoot);
 	}
 
-	private void initModules(SettingsNodeBase oRoot)
+	private void initModules(TreeNodeBase oRoot)
 	{
-		for(SettingsNodeBase node : oRoot.getChildModules())
+		for(TreeNodeBase node : oRoot.getChildModules())
 		{
 			oRoot.add(node);
 			if(oRoot.getChildModules().size() > 0)
