+++ src/crossdev64/settings/nodes/SettingsRootNode.java
@@ -1,6 +1,7 @@
 package crossdev64.settings.nodes;
 
 import crossdev64.emulator.EmulatorModuleNodes;
+import crossdev64.gui.TreeNodeBase;
 import crossdev64.plugins.PluginsNode;
 import crossdev64.settings.GlobalSettings;
 
@@ -10,7 +11,7 @@ import crossdev64.settings.GlobalSettings;
  *
  */
 public class SettingsRootNode
-	extends SettingsNodeBase
+	extends TreeNodeBase
 {
 	public static final String MODULE_ID = "5A75ED5D-6A8E-48F5-A1EF-68C823D8443E";
